Rename module to match new StoragePool class name.
